IRC parsing: enforce nickname to be immediately after timestamp
The nickname always comes immediately after the timestamp (with a
varying number of whitespaces, as observed in e.g. n24098). By
making the nickname regex pattern more strict (i.e. requiring it
to come right after the timestamp), we can avoid issues with
messages that don't contain a username but do contain the < >
brackets, such as in
https://github.com/bitcoin-core-review-club/website/pull/584/files