contrib: bugfix: allow empty user.ghtoken
subprocess.run() does not throw a CalledProcessError when the
process exits with a non-zero exit code, unless `check=True` is
set. This means that when the `option` key is not found, an
empty string is returned instead, causing a wrong ghtoken
to be used for the `Authorization` header.

This leads to:
Github returned error HTTP Error 401: Unauthorized

Fix this and simplify the code by using the convenience
function `subprocess.check_output()` instead.